requirements.txt
progress==1.5
PyGithub==1.52
urllib3==1.25.10
dohq-tfs==1.0.175